Implement scheduler.util.backoff as a queue
We are going to use PodBackoff for controlling backoff when adding
unschedulable pods back to the active scheduling queue. In order to do
this more easily, limit the interface for PodBackoff to only this struct
(rather than exposing BackoffEntry as well) and change the backing
implementation to be a queue which can be used to wait for the next
BackoffEntry completion.

This allows us to use backoff as a priority queue making it easier to
wait for the next backoff completion.